EX-220 - Made major enhancements to the ViewText and ViewBinary classes in the Stream library, along with other minor improvements to the stream library. Also fixed some build system issues and cleaned up some compiler warnings.

+++ Support Libraries/Stream/WAVFile.inl
@@ -319,7 +319,7 @@ bool WAVFile::Open(const std::wstring& filename, OpenMode openMode, CreateMode c
 	{
 		bufferSize = abufferSize;
 		delete[] fileBuffer;
-		fileBuffer = new unsigned char[bufferSize];
+		fileBuffer = new unsigned char[(size_t)bufferSize];
 	}
 	bufferPosOffset = 0;
 	bytesRemainingInBuffer = bufferSize;
